refactor: 拆分独立窗口界面hooks组件和配置文件
